implemented, documented, not tested
