+++ pyccl/emulators/__init__.py
@@ -1,4 +1,4 @@
 from .emu_base import *
 from .baccoemu_linear_pk import *
 from .baccoemu_nonlinear_pk import *
-from .cosmicemu_MTIV_pk import *
+from .cosmicemu_pk import *
