--- readthedocs/api/pyccl.emulators.cosmicemu_MTIV_pk.rst
@@ -1,7 +1,0 @@
-pyccl.emulators.cosmicemu\_MTIV\_pk module
-==========================================
-
-.. automodule:: pyccl.emulators.cosmicemu_MTIV_pk
-   :members:
-   :undoc-members:
-   :show-inheritance:
