+++ readthedocs/api/pyccl.emulators.cosmicemu_pk.rst
@@ -1,0 +1,7 @@
+pyccl.emulators.cosmicemu\_pk module
+====================================
+
+.. automodule:: pyccl.emulators.cosmicemu_pk
+   :members:
+   :undoc-members:
+   :show-inheritance:
